Recover from a JSON parser failure by just treating the value as a string
Submitting via:

    curl ... -F foo=bar

is treated the same way as:

    curl ... -F 'foo="bar"'

This is somewhat undesirable, but avoids us accidentally including
unescaped doublequotes into string values. This sets a general precident
of:

    curl ... -F 'foo=\"bar\"'

being the blessed way to ensure JSON tokens are not accidentally
processed when they were intended to be included in the value.