Represent companion object as its components
Since Java doesn't have the concept of a companion object, we can't use
the `companion` member of a few of the term case classes properly.  The
closest thing Java has is static members on a regular class, so we'll
represent the companion as just a list of imports, declarations, object
members, and values, and wrap them in an object when building Scala
class files.