Don't use DTO's Builder class for Jackson deserialization
Unfortunately it's not going to work with how I want to do things.  For
serialization, it was fine.  For deserialization, Jackson was passing
null to the with*() method on the Builder when particular fields were
null or missing, which was causing Optional.of() to throw an exception.
I don't want to change it to Optional.ofNullable(), because I want to
disallow _humans_ passing null to those methods.