Make Jackson polymorphism actually work properly
The old way didn't work because all POJO classes had private
construtors, and so if a base class had properties, there was no way to
properly subclass it.

This is also a bit of a redesign.  Now each class defn in the hierarchy
"owns" its own properties (that is, it has fields for the properties
declared on it).  Child classes also own their own properties, but are
responsible for also taking care of their parents' properties in the
builder, and passing them to superclasses via super() calls in the
constructors.

In the end, this turns out to be a lot more consistent and easier to
reason about.