Flatten Jackson inheritance hierarchies where needed
This is to prevent us from generating uncompileable multiple-inheritance
hierarchies.  It works like this:

1) If there's more than one parent class that has a discriminator, we
throw an error.

2) If there's a single parent that has a discriminator, we extend from
that parent class.  If there are other parents, we flatten them so the
generated class has the properties of all those other parents.

3) If there's just a single parent, we extend from that parent class.

4) Otherwise there's no composition or polymorphism going on, so we just
emit a class that extends from nothing.

Unfortunately this necessitated removing the static `builder()` method
from the POJO class, because you can't have a static method in a
subclass with the same arg list that has a different return type.
(Which is an odd restriction since Java doesn't support static method
overriding, but perhaps they did it for future-proofing.)  To
compensate, the builder constructors are now public.