Remove AHC baseUrl/clientName final modifier when we don't want it
If the openapi spec gives us a baseUrl or clientName, we want to
initialize it as the default, but allow overriding using a method on the
builder.  If there is none provided, we take it in the constructor and
offer no means to override.

A side note: it looks like FieldDeclaration#setFinal() doesn't really
work.  Odd.