Allow overriding required POJO properties with builder methods
Since required properties are passed to the builder constructor, there's
usually no need to be able to call `withSomeProperty(...)` to later set
the value (again).  However, since I added the "copy" constructor for
the builder, this would mean that you wouldn't be able to use that
constructor and then modify the value of a required property, which
greatly reduces its usefulness.  So, we remove that restriction and add
a few more checks to ensure that the object remains consistent.