Don't generate an inline object for objects with no properties
The core doesn't really know how to emit objects like that (so it
doesn't, and we end up with a compile error when building the generated
classes).  At any rate, if there's a bare object type without any
properties, our assumption is that the user is expected to populate it
with arbitrary key/value pairs (sorta like a map, but I guess they get
more freedom for keys?), so in this case we should just use
objectType() (io.cire.Json or Jackson JsonNode).