feat: add support for http4s AuthedRoutes middleware
http4s include a middleware for authenticating users. To enable this
feature you need to add the `secutity` configuration to the openApi
specification (see server3 for an example) and include the `--http4sAuthedRoutes`
argument when generating the code. Note, this is only the server-side
implementation. The client already has the option to pass headers.

Signed-off-by: Ingar Abrahamsen <ingar.abrahamsen@tapad.com>