Add broader ContentType support possibilities
This adds generic TextContent and BinaryContent classes that represent
arbitrary content-types outside of our main directly-supported types
(like application/json, text/plain, etc.).  The goal here is to
transparently support arbitrary text and binary types.

For example, imagine an API that handles document storage.  The API
would likely make use of JSON for handling metadata about documents, but
there could be endpoints that return the documents themselves.  It would
be nice if we could support, say, application/pdf documents without
having to special case every possible document type.

The classification of text vs. binary when it comes to the application/
registry is a bit fuzzy, since there's no encoding information attached
to the content-type string itself.  There are a few things we can be
sure about: if the type ends in "+json" or "+xml", then it's going to be
text.  I've also hard-coded a few possibly-useful content-types that are
known to be text.  Otherwise, things in application/ are assumed to be
binary.  If there are other types that should be text, and people ask
for them, adding support is a one-line addition.

One wart: I've explicitly listed application/xml in
ContentType.unapply() as unsupported (so unapply() will return None),
because accepting it will break codegen for existing specs that have
(for example) both application/json and application/xml listed as
content types for a particular response.  At some point if we decide to
actually support XML (whether as a first-class citizen, or just passing
it through as a String, or using scala-xml's NodeSeq, etc.), we can add
it back.