Use InputStream instead of File in Dropwizard
... except for multipart/form-data.

InputStream is nearly always a better choice:

For client requests, users might have data in memory.  They shouldn't
have to write it out to a file before sending it.  If they _do_ already
have a file, then they can wrap it in a FileInputStream.  If they have a
byte array, they can wrap it in ByteArrayInputStream.  And so on...

For client responses, users usually won't want us to write files out to
disk when we receive binary data.  In the rare case that they do, they
can do so themselves.

On the server request side, we don't want to require that the server
write the entire request body out to disk before starting up handler
logic, and if the user is just going to examine/parse that data in
memory, or ship it off to another network location, writing to a file is
wasteful.  If the user really does want a file, they can write it out.

On the server response side, it's probably rare that a response body is
going to be sitting in a file on disk.  And even when it is, that's
doable with a FileInputStream.

The _one_ case where we're going to continue to use File is for
multipart/form-data, because the only way for the server to properly
parse the input and get to the point where it can call the resource
method is for it to read in the _entire_ request body, which means it'll
have to buffer the entire thing in memory if we use InputStream.  In
this case, using File is a better choice.