Add a Jackson Scala protocol generator
This is composed on top of the circe generator; it runs the circe
generator just for structure, swaps out decoder/encoder, and adds
Jackson annotations to the classes and method parameters where needed.

This is phase one: phase two will disentangle some of the expectation in
the akka-http (etc.) framework generators that circe is being used.