Widen JavaLanguage#Statement to Node, widen RenderedRoutes.routes
This is somewhat unfortunate, but scalameta and javaparser model their
AST hierarchy a bit differently:

Scalameta considers most things a Stat, including Term.  In generic
parlance, scalameta considers an "expression" to be a type of
"statement".

In contrast, javaparser has a root interface, Node, from which Statement
and Expression inherit, independently.

This matters here because akka-http uses a combined Term for its route
specification (since akka-http routes are just expressions).  But in
dropwizard, route specifications are method definitions.

For java-dropwizard, this worked "fine", because JavaLanguage#Term is
defined as Node (because we need that to allow both expressions and
statements).  But for scala-dropwizard this is not, because
ScalaLanguage#Term is just Term, which does not cover method definitions
(Defn.Def, which are Stats, but not Terms).

Ultimately I think maybe I should refactor LanguageAbstraction a bit so
we can have three types:

* "node"
  * scala: Stat
  * java: Node
* "statement"
  * scala: Stat
  * java: Statement
* "expression"
  * scala: Term
  * java Expression

I *think* that would work ok, though I'm not 100% sure it's reasonable
to restrict "expression" to Term in scala.  Anyhow, that's for another
day.