Actually fix DW resource optional collection param unwrapping
Looks like I messed it up when I implemented this fix before; added a
test this time to make sure it works.

The background is that Jersey can't support nested collections types as
resource method parameters.  So Optional<List<T>> does not work.
Instead, we want to just unwrap that as List<T>, and Jersey will give us
an empty list if it's not provided.