WIP: Move to Dropwizard 2.0 series
Why:
1.3 branch was last released in 2020-12 making it pretty stale. Moving
off of this branch keeps us with the releases.

How:
This should remove all references to dropwizard 1.3.X releases and
should add the specific dependencies required for this
project. Specifically, with Dropwizard 2.X, Guardrail needs to include

1. `hibernate-core`
2. `hibernate-validator`

as they are no longer transitively pulled into the project via
Dropwizard dependencies.

Need Help:
I don't understand how the classpath for the tests is
constructed. This results in the classpath missing the `hibernate`
dependencies during test runs.