When releasing, explicitly depend on published jar
This will make builds more stable at the expense of pain during release.

If the pain is great enough, it may be sufficient to do a publishLocal
for dependencies from known tags instead of waiting for a repository
sync.

Initially we were just faking it by pretending that master had the same
ABI as releases because the release script triggers new releases in
order, but this was always a trick, albeit one that I believed would
work.